missed this ProjectShowcase

+++ src/DevelopmentHell.Hubba/ProjectShowcaseManager/Abstractions/IProjectShowcaseManager.cs
@@ -29,7 +29,7 @@ namespace DevelopmentHell.Hubba.ProjectShowcase.Manager.Abstractions
         Task<Result<List<ShowcaseComment>>> GetComments(string showcaseId, int? commentCount, int? page);
         Task<Result<List<CommentReport>>> GetAllCommentReports();
         Task<Result<List<CommentReport>>> GetCommentReports(int commentId);
-        Task<Result<float>> LikeShowcase(string showcaseId);
+        Task<Result<double>> LikeShowcase(string showcaseId);
         Task<Result<string>> CreateShowcase(int listingId, string title, string description, List<IFormFile> files);
         Task<Result> EditShowcase(string showcaseId, int? listingId, string? title, string? description, List<IFormFile>? files);
         Task<Result> DeleteShowcase(string showcaseId);
